Fix validator namespace
