+++ app/validators/deferred/base_validator.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-module Deferral
+module Deferred
   class BaseValidator < ActiveModel::EachValidator
     def validate_each(record, attribute, value)
       record.errors.add(attribute, I18n.t("errors.defer_reason.invalid")) unless valid?(value)
