+++ app/validators/deferred/ecf_validator.rb
@@ -1,13 +1,16 @@
 # frozen_string_literal: true
 
-module Deferral
-  class NPQValidator < BaseValidator
+module Deferred
+  class ECFValidator < BaseValidator
     class << self
       def reasons
         %w[
           left-teaching-profession
           moved-school
+          mentor-no-longer-being-mentor
+          school-left-fip
           career-break
+          passed-induction
           other
         ].freeze
       end
