+++ app/validators/deferred/npq_validator.rb
@@ -1,16 +1,13 @@
 # frozen_string_literal: true
 
-module Deferral
-  class ECFValidator < BaseValidator
+module Deferred
+  class NPQValidator < BaseValidator
     class << self
       def reasons
         %w[
           left-teaching-profession
           moved-school
-          mentor-no-longer-being-mentor
-          school-left-fip
           career-break
-          passed-induction
           other
         ].freeze
       end
